Fixing after rebase
Moved MS-specific object types and vocabularies to their own subfolder

Fixed named from MS_MASSRANGE to MS_MASS_RANGE
