+++ bam_masterdata/datamodel/mass_spectrometry/__init__.py
@@ -1,0 +1,1 @@
+"""Submodule containing type definitions related to mass spectrometry (MS) experiments."""
