+++ bam_masterdata/datamodel/object_types.py
@@ -7525,111 +7525,6 @@ class Freezer2(Instrument):
     )
 
 
-class MassSpec(Instrument):
-    defs = ObjectTypeDef(
-        code="INSTRUMENT.MASS_SPEC",
-        description="""Mass Spectrometer//Massenspektrometer""",
-        generated_code_prefix="INS.MS",
-    )
-
-    mass_spec_type = PropertyTypeAssignment(
-        code="MASS_SPEC_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="MASS_SPEC_TYPE",
-        property_label="MS Type",
-        description="""Mass Spectrometer Type//Massenspektrometer-Typ""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-    # changes
-    ion_source = PropertyTypeAssignment(
-        code="IONIZATION_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="IONIZATION_TYPE",
-        property_label="Ion source",
-        description="""Ionization Type//Ionenquelle""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-    chromatography = PropertyTypeAssignment(
-        code="CHROMATOGRAPHY_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="CHROMATOGRAPHY_TYPE",
-        property_label="Chromatography",
-        description="""Separator Type//Trennverfahren""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-    detector = PropertyTypeAssignment(
-        code="DETECTOR_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="DETECTOR_TYPE",
-        property_label="Detector",
-        description="""Detector Type//Analysator""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-
-class LcSystem(Instrument):
-    defs = ObjectTypeDef(
-        code="INSTRUMENT.LC_SYSTEM",
-        description="""LC-System//LC-System""",
-        generated_code_prefix="INS.LC",
-    )
-
-    detector_type = PropertyTypeAssignment(
-        code="DETECTOR_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="DETECTOR_TYPE",
-        property_label="Detector",
-        description="""Detector Type//Analysator""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-    detector_type_secondary = PropertyTypeAssignment(
-        code="DETECTOR_TYPE_SECONDARY",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="DETECTOR_TYPE",
-        property_label="Detector",
-        description="""Detector Type//Analysator""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-
-class GcSystem(Instrument):
-    defs = ObjectTypeDef(
-        code="INSTRUMENT.GC_SYSTEM",
-        description="""GC-System//GC-System""",
-        generated_code_prefix="INS.GC",
-    )
-
-    detector_type = PropertyTypeAssignment(
-        code="DETECTOR_TYPE",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="DETECTOR_TYPE",
-        property_label="Detector",
-        description="""Detector Type//Analysator""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-    detector_type_secondary = PropertyTypeAssignment(
-        code="DETECTOR_TYPE_SECONDARY",
-        data_type="CONTROLLEDVOCABULARY",
-        vocabulary_code="DETECTOR_TYPE",
-        property_label="Detector",
-        description="""Detector Type//Analysator""",
-        mandatory=False,
-        section="Technical Details",
-    )
-
-
 class Scale(Instrument):
     defs = ObjectTypeDef(
         code="INSTRUMENT.SCALE",
